Test1

+++ storage/postgres/book.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"database/sql"
 	"fmt"
+
 	"github.com/OybekAbduvosiqov/book/models"
 
 	"github.com/google/uuid"
@@ -257,9 +258,22 @@ func (r *BookRepo) Update(ctx context.Context, book *models.UpdateBook) error {
 }
 
 func (r *BookRepo) Delete(ctx context.Context, req *models.BookPrimeryKey) error {
-	_, err := r.db.Exec(ctx, "DELETE FROM book_category WHERE books_id  = $1 ", req.Id)
+
+	var (
+		count int
+	)
+
+	err := r.db.QueryRow(ctx, "SELECT COUNT(*) FROM book_category WHERE books_id = $1", req.Id).Scan(&count)
 	if err != nil {
 		return err
+	}
+
+	if count > 0 {
+		_, err := r.db.Exec(ctx, "DELETE FROM book_category WHERE books_id  = $1 ", req.Id)
+		if err != nil {
+			return err
+		}
+
 	}
 	_, err = r.db.Exec(ctx, "DELETE FROM books WHERE id = $1", req.Id)
